fix - include changes missed from last commit

+++ src/netbsd/mod.rs
@@ -323,8 +323,8 @@ impl GeneralReadout for NetBSDGeneralReadout {
             let disk_size = stats.f_blocks * stats.f_bsize as u64;
             let free = stats.f_bavail * stats.f_bsize as u64;
 
-            let used_byte = (disk_size - free) as u128;
-            let disk_size_byte = disk_size as u128;
+            let used_byte = (disk_size - free);
+            let disk_size_byte = disk_size;
 
             return Ok((used_byte, disk_size_byte));
         }
